"Upgrade" Web3 to 1.3.4
Rework significant portions of the code in order to use `web3@1.3.4` as
a base instead of `web3@1.0.0-beta.55`. The versioning story for Web3 is
long and I won't detail it here: in short, `1.3.4` and `1.0.0-beta.55`
come from different branches where the `1.0.0-beta` branch is in a sense
"newer" than `1.3.4`, although it appears to have less support. For
details, see
https://medium.com/@samuel_91690/1-0-release-web3-js-ddd23d3c8f62.

Converting the code to `web3@1.3.4` requires significant changes. Most
notably, the interface for providers, especially those that support
subscriptions, is totally different. Further, the `subscriptionFactory`
we used to add custom subscription types has been removed, so we
implement custom subscriptions in another way.